pyproject.toml
[project]
name = "expression-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "einops>=0.8.1",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
